// src/types/jss.ts
import {
	MARGIN,
	MARGIN_BOTTOM,
	MARGIN_LEFT,
	MARGIN_RIGHT,
	MARGIN_TOP,
	PADDING,
	PADDING_BOTTOM,
	PADDING_LEFT,
	PADDING_RIGHT,
	PADDING_TOP,
} from "css-properties-props";

export interface MarginProps {
	margin?: MARGIN;
	marginTop?: MARGIN_TOP;
	marginBottom?: MARGIN_BOTTOM;
	marginLeft?: MARGIN_LEFT;
	marginRight?: MARGIN_RIGHT;
	marginXaxis?: string;
	marginYaxis?: string;

	//   Xs
	marginXs?: MARGIN;
	marginXsTop?: MARGIN_TOP;
	marginXsBottom?: MARGIN_BOTTOM;
	marginXsLeft?: MARGIN_LEFT;
	marginXsRight?: MARGIN_RIGHT;
	marginXsXaxis?: string;
	marginXsYaxis?: string;

	//   Sm
	marginSm?: MARGIN;
	marginSmTop?: MARGIN_TOP;
	marginSmBottom?: MARGIN_BOTTOM;
	marginSmLeft?: MARGIN_LEFT;
	marginSmRight?: MARGIN_RIGHT;
	marginSmXaxis?: string;
	marginSmYaxis?: string;

	//   Md
	marginMd?: MARGIN;
	marginMdTop?: MARGIN_TOP;
	marginMdBottom?: MARGIN_BOTTOM;
	marginMdLeft?: MARGIN_LEFT;
	marginMdRight?: MARGIN_RIGHT;
	marginMdXaxis?: string;
	marginMdYaxis?: string;

	//   Lg
	marginLg?: MARGIN;
	marginLgTop?: MARGIN_TOP;
	marginLgBottom?: MARGIN_BOTTOM;
	marginLgLeft?: MARGIN_LEFT;
	marginLgRight?: MARGIN_RIGHT;
	marginLgXaxis?: string;
	marginLgYaxis?: string;

	//   Xl
	marginXl?: MARGIN;
	marginXlTop?: MARGIN_TOP;
	marginXlBottom?: MARGIN_BOTTOM;
	marginXlLeft?: MARGIN_LEFT;
	marginXlRight?: MARGIN_RIGHT;
	marginXlXaxis?: string;
	marginXlYaxis?: string;
}

export interface PaddingProps {
	padding?: PADDING;
	paddingTop?: PADDING_TOP;
	paddingBottom?: PADDING_BOTTOM;
	paddingLeft?: PADDING_LEFT;
	paddingRight?: PADDING_RIGHT;
	paddingXaxis?: string;
	paddingYaxis?: string;

	//   Xs
	paddingXs?: PADDING;
	paddingXsTop?: PADDING_TOP;
	paddingXsBottom?: PADDING_BOTTOM;
	paddingXsLeft?: PADDING_LEFT;
	paddingXsRight?: PADDING_RIGHT;
	paddingXsXaxis?: string;
	paddingXsYaxis?: string;

	//   Sm
	paddingSm?: PADDING;
	paddingSmTop?: PADDING_TOP;
	paddingSmBottom?: PADDING_BOTTOM;
	paddingSmLeft?: PADDING_LEFT;
	paddingSmRight?: PADDING_RIGHT;
	paddingSmXaxis?: string;
	paddingSmYaxis?: string;

	//   Md
	paddingMd?: PADDING;
	paddingMdTop?: PADDING_TOP;
	paddingMdBottom?: PADDING_BOTTOM;
	paddingMdLeft?: PADDING_LEFT;
	paddingMdRight?: PADDING_RIGHT;
	paddingMdXaxis?: string;
	paddingMdYaxis?: string;

	//   Lg
	paddingLg?: PADDING;
	paddingLgTop?: PADDING_TOP;
	paddingLgBottom?: PADDING_BOTTOM;
	paddingLgLeft?: PADDING_LEFT;
	paddingLgRight?: PADDING_RIGHT;
	paddingLgXaxis?: string;
	paddingLgYaxis?: string;

	//   Xl
	paddingXl?: PADDING;
	paddingXlTop?: PADDING_TOP;
	paddingXlBottom?: PADDING_BOTTOM;
	paddingXlLeft?: PADDING_LEFT;
	paddingXlRight?: PADDING_RIGHT;
	paddingXlXaxis?: string;
	paddingXlYaxis?: string;
}

export interface ColProps {
	col?: 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;
	xs?: 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;
	sm?: 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;
	md?: 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;
	lg?: 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;
	xl?: 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;
}